### 2025-07-10 #8431 @kthyng
@cjtu I am going to ping our editorial board to see if this submission passes our criteria for [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This process will take a few weeks; thanks for your patience.
