### 2025-07-29 #8663 @editorialbot
[REVIEW]: Craterpy: Impact crater data science in Python
**Submitting author:** @cjtu (<a href="http://orcid.org/0000-0001-9972-1534">Christian J. Tai Udovicic</a>)
**Repository:** https://github.com/cjtu/craterpy
**Branch with paper.md** (empty if default branch):
**Version:** v0.9.1
