### 2025-08-09 cjtu/craterpy#84 @athulpg007
Consider adding a GitHub Action to run the tests on pull request
Please consider adding a GitHub action to run the tests in https://github.com/cjtu/craterpy/tree/main/craterpy/tests on pull requests.
