### 2025-08-09 cjtu/craterpy#85 @athulpg007
[optional] Consider using uv to install packages
The package has a reasonably large number of dependencies. The docs suggest using poetry to set up a dev environment, Consider exploring [uv](https://docs.astral.sh/uv/) instead of poetry as the package manager, as it may provide a considerable performance improvement while setting up the env. This is optional and not required.
```
Package operations: 132 installs, 0 updates, 0 removals

  - Installing attrs (25.3.0)
  - Installing rpds-py (0.24.0)
  - Installing typing-extensions (4.13.1)
  - Installing referencing (0.36.2)
  - Installing six (1.17.0)
  - Installing jsonschema-specifications (2024.10.1)
  - Installing traitlets (5.14.3)
  - Installing python-dateutil (2.9.0.post0)
  - Installing platformdirs (4.3.7)
  - Installing types-python-dateutil (2.9.0.20241206)
  - Installing arrow (1.3.0)
  - Installing fastjsonschema (2.21.1)
  - Installing jsonschema (4.23.0)
  - Installing jupyter-core (5.7.2)
  - Installing pycparser (2.22)
  - Installing pyzmq (26.4.0)
  - Installing tornado (6.4.2)
  - Installing webencodings (0.5.1)
  - Installing cffi (1.17.1)
  - Installing fqdn (1.5.1)
  - Installing idna (3.10)
  - Installing isoduration (20.11.0)
  - Installing jsonpointer (3.0.0)
  - Installing jupyter-client (8.6.3)
  - Installing markupsafe (3.0.2)
  - Installing nbformat (5.10.4)
  - Installing ptyprocess (0.7.0)
  - Installing rfc3339-validator (0.1.4)
  - Installing rfc3986-validator (0.1.1)
  - Installing soupsieve (2.6)
  - Installing tinycss2 (1.4.0)
  - Installing uri-template (1.3.0)
  - Installing webcolors (24.11.1)
  - Installing argon2-cffi-bindings (21.2.0): Downloading... 80%
  - Installing asttokens (3.0.0)
  - Installing beautifulsoup4 (4.13.3): Downloading... 50%
  - Installing asttokens (3.0.0)
  - Installing beautifulsoup4 (4.13.3): Downloading... 50%
  - Installing argon2-cffi-bindings (21.2.0): Downloading... 100%
  - Installing asttokens (3.0.0)
  - Installing beautifulsoup4 (4.13.3): Downloading... 50%
  - Installing bleach (6.2.0): Downloading... 70%
  - Installing bleach (6.2.0): Downloading... 70%
  - Installing beautifulsoup4 (4.13.3): Downloading... 61%
  - Installing bleach (6.2.0): Downloading... 70%
  - Installing bleach (6.2.0): Downloading... 90%
  - Installing asttokens (3.0.0)
  - Installing beautifulsoup4 (4.13.3): Downloading... 61%
  - Installing bleach (6.2.0): Downloading... 90%
  - Installing argon2-cffi-bindings (21.2.0): Installing...
  - Installing asttokens (3.0.0)
  - Installing beautifulsoup4 (4.13.3): Downloading... 61%
  - Installing bleach (6.2.0): Downloading... 90%
  - Installing bleach (6.2.0): Downloading... 90%
  - Installing beautifulsoup4 (4.13.3): Downloading... 81%
  - Installing bleach (6.2.0): Downloading... 90%
  - Installing bleach (6.2.0): Downloading... 100%
  - Installing bleach (6.2.0): Downloading... 100%
  - Installing beautifulsoup4 (4.13.3): Downloading... 100%
  - Installing bleach (6.2.0): Downloading... 100%
  - Installing bleach (6.2.0): Installing...
  - Installing asttokens (3.0.0)
  - Installing beautifulsoup4 (4.13.3): Downloading... 100%
  - Installing bleach (6.2.0): Installing...
  - Installing argon2-cffi-bindings (21.2.0)
  - Installing asttokens (3.0.0)
  - Installing beautifulsoup4 (4.13.3): Downloading... 100%
  - Installing bleach (6.2.0): Installing...
  - Installing defusedxml (0.7.1)
  - Installing bleach (6.2.0): Installing...
  - Installing defusedxml (0.7.1)
  - Installing beautifulsoup4 (4.13.3): Installing...
  - Installing bleach (6.2.0): Installing...
  - Installing defusedxml (0.7.1)
  - Installing defusedxml (0.7.1)
  - Installing bleach (6.2.0)
  - Installing defusedxml (0.7.1)
  - Installing bleach (6.2.0)
  - Installing defusedxml (0.7.1)
  - Installing beautifulsoup4 (4.13.3)
  - Installing bleach (6.2.0)
  - Installing defusedxml (0.7.1)
  - Installing executing (2.2.0)
  - Installing jinja2 (3.1.6)
  - Installing jupyterlab-pygments (0.3.0)
  - Installing mistune (3.1.3)
  - Installing nbclient (0.10.2)
  - Installing packaging (24.2)
  - Installing pandocfilters (1.5.1)
  - Installing parso (0.8.4)
  - Installing pure-eval (0.2.3)
  - Installing pygments (2.19.1)
  - Installing pyyaml (6.0.2)
  - Installing sniffio (1.3.1)
  - Installing python-json-logger (3.3.0)
  - Installing terminado (0.18.1)
  - Installing wcwidth (0.2.13)
  - Installing anyio (4.9.0): Downloading... 81%
  - Installing argon2-cffi (23.1.0)
  - Installing certifi (2025.1.31): Downloading... 51%
  - Installing argon2-cffi (23.1.0)
  - Installing certifi (2025.1.31): Downloading... 51%
  - Installing anyio (4.9.0): Downloading... 100%
  - Installing argon2-cffi (23.1.0)
  - Installing certifi (2025.1.31): Downloading... 51%
  - Installing certifi (2025.1.31): Downloading... 61%
  - Installing charset-normalizer (3.4.1): Downloading... 62%
  - Installing charset-normalizer (3.4.1): Downloading... 81%
  - Installing charset-normalizer (3.4.1): Downloading... 81%
  - Installing certifi (2025.1.31): Downloading... 71%
  - Installing charset-normalizer (3.4.1): Downloading... 81%
  - Installing argon2-cffi (23.1.0)
  - Installing certifi (2025.1.31): Downloading... 71%
  - Installing charset-normalizer (3.4.1): Downloading... 81%
  - Installing anyio (4.9.0): Installing...
  - Installing argon2-cffi (23.1.0)
  - Installing certifi (2025.1.31): Downloading... 71%
  - Installing charset-normalizer (3.4.1): Downloading... 81%
  - Installing charset-normalizer (3.4.1): Downloading... 90%
  - Installing charset-normalizer (3.4.1): Downloading... 90%
  - Installing certifi (2025.1.31): Downloading... 81%
  - Installing charset-normalizer (3.4.1): Downloading... 90%
  - Installing charset-normalizer (3.4.1): Downloading... 100%
  - Installing charset-normalizer (3.4.1): Downloading... 100%
  - Installing certifi (2025.1.31): Downloading... 100%
  - Installing charset-normalizer (3.4.1): Downloading... 100%
  - Installing charset-normalizer (3.4.1): Installing...
  - Installing decorator (5.2.1)
  - Installing charset-normalizer (3.4.1): Installing...
  - Installing decorator (5.2.1)
  - Installing certifi (2025.1.31): Installing...
  - Installing charset-normalizer (3.4.1): Installing...
  - Installing decorator (5.2.1)
  - Installing argon2-cffi (23.1.0)
  - Installing certifi (2025.1.31): Installing...
  - Installing charset-normalizer (3.4.1): Installing...
  - Installing decorator (5.2.1)
  - Installing anyio (4.9.0)
  - Installing argon2-cffi (23.1.0)
  - Installing certifi (2025.1.31): Installing...
  - Installing charset-normalizer (3.4.1): Installing...
  - Installing decorator (5.2.1)
  - Installing charset-normalizer (3.4.1): Installing...
  - Installing decorator (5.2.1)
  - Installing certifi (2025.1.31)
  - Installing charset-normalizer (3.4.1): Installing...
  - Installing decorator (5.2.1)
  - Installing decorator (5.2.1)
  - Installing charset-normalizer (3.4.1)
  - Installing decorator (5.2.1)
  - Installing h11 (0.14.0)
  - Installing ipython-pygments-lexers (1.1.1)
  - Installing jedi (0.19.2)
  - Installing jupyter-events (0.12.0)
  - Installing jupyter-server-terminals (0.5.3)
  - Installing matplotlib-inline (0.1.7)
  - Installing nbconvert (7.16.6)
  - Installing overrides (7.7.0)
  - Installing pexpect (4.9.0)
  - Installing prometheus-client (0.21.1)
  - Installing prompt-toolkit (3.0.50)
  - Installing send2trash (1.8.3)
  - Installing stack-data (0.6.3)
  - Installing urllib3 (2.3.0)
  - Installing websocket-client (1.8.0)
  - Installing babel (2.17.0)
  - Installing comm (0.2.2)
  - Installing debugpy (1.8.13)
  - Installing httpcore (1.0.7)
  - Installing ipython (9.0.2)
  - Installing json5 (0.12.0)
  - Installing jupyter-server (2.15.0)
  - Installing nest-asyncio (1.6.0)
  - Installing psutil (7.0.0)
  - Installing requests (2.32.3)
  - Installing async-lru (2.0.5)
  - Installing click (8.1.8)
  - Installing httpx (0.28.1)
  - Installing ipykernel (6.29.5)
  - Installing jupyter-lsp (2.2.5)
  - Installing jupyterlab-server (2.27.3)
  - Installing notebook-shim (0.2.4)
  - Installing numpy (2.2.4)
  - Installing setuptools (78.1.0)
  - Installing affine (2.4.0)
  - Installing click-plugins (1.1.1)
  - Installing cligj (0.7.2)
  - Installing contourpy (1.3.1)
  - Installing cycler (0.12.1)
  - Installing fonttools (4.57.0)
  - Installing iniconfig (2.1.0)
  - Installing jupyterlab (4.3.6)
  - Installing jupyterlab-widgets (3.0.13)
  - Installing kiwisolver (1.4.8)
  - Installing pillow (11.1.0)
  - Installing pluggy (1.5.0)
  - Installing pyparsing (3.2.3)
  - Installing pytz (2025.2)
  - Installing tzdata (2025.2)
  - Installing widgetsnbextension (4.0.13)
  - Installing coverage (7.8.0): Downloading... 40%
  - Installing coverage (7.8.0): Downloading... 60%
  - Installing coverage (7.8.0): Downloading... 80%
  - Installing coverage (7.8.0): Downloading... 100%
  - Installing coverage (7.8.0): Installing...
  - Installing coverage (7.8.0)
  - Installing fiona (1.10.1)
  - Installing ipywidgets (8.1.5)
  - Installing jupyter-console (6.6.3)
  - Installing matplotlib (3.10.1)
  - Installing mypy-extensions (1.0.0)
  - Installing notebook (7.3.3)
  - Installing pandas (2.2.3)
  - Installing pathspec (0.12.1)
  - Installing pyogrio (0.10.0)
  - Installing pyproj (3.7.1)
  - Installing pyshp (2.3.1)
  - Installing pytest (8.3.5)
  - Installing rasterio (1.4.3)
  - Installing shapely (2.1.0)
  - Installing simplejson (3.20.1)
  - Installing antimeridian (0.4.0)
  - Installing black (25.1.0)
  - Installing cartopy (0.24.1)
  - Installing geopandas (1.0.1)
  - Installing jupyter (1.1.1)
  - Installing pytest-cov (6.1.1)
  - Installing rasterstats (0.20.0)
  - Installing scipy (1.15.2)

```
