### 2025-08-09 cjtu/craterpy#86 @athulpg007
Images not rendering in the docs home page
The images appear to not render in the docs home page: [https://craterpy.readthedocs.io/en/latest/index.html](https://craterpy.readthedocs.io/en/latest/index.html)

See "Vesta map plot", "Vesta plot rois"
