### 2025-08-09 cjtu/craterpy#87 @athulpg007
Error when running example
When running the example here: [https://craterpy.readthedocs.io/en/latest/getting_started.html](https://craterpy.readthedocs.io/en/latest/getting_started.html)
```
from craterpy import CraterDatabase
cdb = CraterpyDatabase("craterpy/data/moon_craters.csv", body="Moon", units="km")

ImportError                               Traceback (most recent call last)
Cell In[2], line 1
----> 1 from craterpy import CraterpyDatabase
      2 cdb = CraterpyDatabase("craterpy/data/moon_craters.csv", body="Moon", units="km")

ImportError: cannot import name 'CraterpyDatabase' from 'craterpy' (/home/athul/Documents/craterpy/craterpy/__init__.py)
```
This appears to be a typo as there is no `CraterpyDatabase`, perhaps referring to `CraterDatabase`.
