### 2025-08-09 #8663 @athulpg007
@cjtu, I have done an initial review and opened these issues below:

- https://github.com/cjtu/craterpy/issues/84
- https://github.com/cjtu/craterpy/issues/85
- https://github.com/cjtu/craterpy/issues/86
- https://github.com/cjtu/craterpy/issues/87
