### 2025-08-11 d98b932
Fix typo and broken images in docs (#86, #87).
