### 2025-08-11 8cafcb8
Replace poetry with uv
