+++ .github/workflows/pypi_publish.yml
@@ -3,24 +3,27 @@ name: PyPI publish
 
 on:
   release:
-    types: [released]
+    types: ["released"]
 
 jobs:
-  build:
-
+  run:
+    name: "Build and publish release"
     runs-on: ubuntu-latest
 
     steps:
-      - uses: actions/checkout@v2
-      - name: Set up Python 3.11
-        uses: actions/setup-python@v4
+      - uses: actions/checkout@v4
+
+      - name: Install uv
+        uses: astral-sh/setup-uv@v3
         with:
-          python-version: 3.11
-      - name: Install poetry
-        run: pip install poetry
-      - name: Bump version number
-        run: poetry version ${{ github.event.release.tag_name }}
-      - name: Build package
-        run: poetry build
-      - name: Publish package
-        run: poetry publish -u __token__ -p ${{ secrets.PYPI_TOKEN }}
+          enable-cache: true
+          cache-dependency-glob: uv.lock
+
+      - name: Set up Python
+        run: uv python install 3.12
+
+      - name: Build
+        run: uv build
+
+      - name: Publish
+        run: uv publish -t ${{ secrets.PYPI_TOKEN }}
