+++ pyproject.toml
@@ -1,98 +1,96 @@
-[tool.poetry]
+[project]
 name = "craterpy"
-version = "0.9.1"
-homepage = "https://github.com/cjtu/craterpy"
+version = "0.9.3"
 description = "Impact crater data science in Python."
-authors = ["Christian J. Tai Udovicic <cj.taiudovicic@gmail.com>"]
-license = "MIT"
 readme = "README.md"
-repository = "https://github.com/cjtu/craterpy"
-documentation = "https://readthedocs.org/projects/craterpy/"
-classifiers=[
-    'Development Status :: 4 - Beta',
-    'Intended Audience :: Science/Research',
-    'License :: OSI Approved :: MIT License',
-    'Natural Language :: English',
-    'Operating System :: OS Independent',
-    'Topic :: Scientific/Engineering :: Astronomy',
-    'Programming Language :: Python :: 3.10',
+authors = [
+    {name = "Christian J. Tai Udovicic", email = "cj.taiudovicic@gmail.com"},
+]
+license = {text = "MIT"}
+requires-python = "<4.0,>=3.10"
+dependencies = [
+    "numpy<3.0.0,>=2.2.4",
+    "matplotlib<4.0.0,>=3.10.1",
+    "pandas<3.0.0,>=2.2.3",
+    "rasterio<2.0.0,>=1.3.10",
+    "cartopy<1.0.0,>=0.24.0",
+    "rasterstats<1.0.0,>=0.20.0",
+    "geopandas<2.0.0,>=1.0.1",
+    "scipy<2.0.0,>=1.13.1",
+    "antimeridian<1.0.0,>=0.4.0",
+]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Science/Research",
+    "License :: OSI Approved :: MIT License",
+    "Natural Language :: English",
+    "Operating System :: OS Independent",
+    "Topic :: Scientific/Engineering :: Astronomy",
+    "Programming Language :: Python :: 3.10",
 ]
 keywords = [
-    'python', 
-    'crater',
-    'data analysis',
-    'planetary',
-    'moon',
-]
-include = [
-    "LICENSE.txt",
+    "python",
+    "crater",
+    "data analysis",
+    "planetary",
+    "moon",
 ]
 
-[tool.poetry.urls]
+[project.urls]
 "Bug Tracker" = "https://github.com/cjtu/craterpy/issues"
+homepage = "https://github.com/cjtu/craterpy"
+repository = "https://github.com/cjtu/craterpy"
+documentation = "https://readthedocs.org/projects/craterpy/"
 
-[tool.poetry.dependencies]
-python = "^3.10,<4.0"
-numpy = "^2.2.4"
-matplotlib = "^3.10.1"
-pandas = "^2.2.3"
-rasterio = "^1.3.10"
-cartopy = "^0.24.0"
-rasterstats = "^0.20.0"
-geopandas = "^1.0.1"
-scipy = "^1.13.1"
-antimeridian = "^0.4.0"
-
-[tool.poetry.group.dev.dependencies]
-pytest = "^8.3.5"
-jupyter = "^1.1.1"
-black = "^25.1.0"
-pytest-cov = "^6.0.0"
-
-[tool.poetry.group.docs]
-optional = true
-
-[tool.poetry.group.docs.dependencies]
-myst-parser = "^1.0.0"
-sphinx = "^6.2.1"
-sphinx-rtd-theme = "^2.0.0"
-sphinx-autobuild = "^2024.10.3"
+[dependency-groups]
+dev = [
+    "pytest<9.0.0,>=8.3.5",
+    "jupyter<2.0.0,>=1.1.1",
[…]
