+++ .readthedocs.yml
@@ -4,7 +4,7 @@ sphinx:
    configuration: docs/source/conf.py
 
 build:
-   os: ubuntu-latest
+   os: ubuntu-24.04
    tools:
       python: "3.13"
    jobs:
