### 2025-08-16 cjtu/craterpy#89 @athulpg007
[optional] Consider adding a parameter to allow the plot to be saved
I assume a common use case is to generate an crater map using the package, and save it (.png, .pdf etc,), for use in a publication. If this is already not possible, consider adding a functionality to the CraterDatabase.plot() to allow the user to save the image.

```
from craterpy import CraterDatabase, sample_data

cdb = CraterDatabase(sample_data['vesta_craters.csv'], 'Vesta', units='m')
cdb.plot(sample_data['vesta.tif'], alpha=0.5, color='tab:green')
```
This shows the image in a notebook, but cannot be saved to a figure.
