### 2025-08-16 #8663 @athulpg007
@cjtu Thanks for addressing my comments.

opened one more issue:
- https://github.com/cjtu/craterpy/issues/89

paper:
- suggest to remove the "." at the end of the paper title (Craterpy: Impact crater data science in Python.)
- suggest to include one representative image (eg: [https://craterpy.readthedocs.io/latest/index.html#demo](https://craterpy.readthedocs.io/latest/index.html#demo)) in the paper for illustrating the capability of the package.
