### 2025-08-19 0266d43
Add savefig option to plot method (closes #89)
