### 2025-08-19 61c7658
Add savefig option to plot method (closes #89) (#90)
