### 2025-08-19 cjtu/craterpy#89 @cjtu
Great suggestion, this is now implemented
