### 2025-08-19 01fd9c2
Add figure to JOSS paper
