### 2025-08-19 f563d3f
Remove tests from coverage
