### 2025-08-19 b810eca
Improve test coverage
