+++ craterpy/tests/test_classes.py
@@ -8,6 +8,7 @@ import json
 import os
 import pyproj
 from pyproj import CRS
+from pyogrio.errors import DataSourceError
 import pandas as pd
 import geopandas as gpd
 import numpy as np
@@ -24,6 +25,7 @@ class TestCraterDatabase(unittest.TestCase):
 
     def setUp(self):
         self.moon_tif = craterpy.sample_data["moon.tif"]
+        self.moon_dem = craterpy.sample_data["moon_dem.tif"]
         self.moon_craters = craterpy.sample_data["moon_craters.csv"]
         self.vesta_tif = craterpy.sample_data["vesta.tif"]
         self.vesta_craters = craterpy.sample_data["vesta_craters.csv"]
@@ -53,12 +55,18 @@ class TestCraterDatabase(unittest.TestCase):
         """Test getting statistics on a region for a raster."""
         cdb = self.moon_cdb.copy()
         cdb.add_annuli("rim", 1, 1.1)
-        stats = cdb.get_stats(self.moon_tif, "rim", ["count"])
-        self.assertIn("count_rim", stats.columns)
+        stats = cdb._get_stats(self.moon_tif, "rim", suffix="test")
+        self.assertIn("count_test", stats.columns)
 
     def test_get_stats_parallel(self):
         """Test parallellization of get_stats for multiple rasters/regions."""
-        pass
+        cdb = self.moon_cdb.copy()
+        cdb.add_annuli("rim", 1, 1.1)
+        stats = cdb.get_stats(
+            {"moon": self.moon_tif, "dem": self.moon_dem}, "rim", ["count"]
+        )
+        self.assertIn("count_moon_rim", stats.columns)
+        self.assertIn("count_dem_rim", stats.columns)
 
     def test_body_crs_all(self):
         """Test that every defined CRS loads."""
@@ -116,21 +124,103 @@ class TestCraterDatabase(unittest.TestCase):
             cdb = CraterDatabase(test_df, "moon")
             self.assertEqual(cdb.rad.iloc[0], 10.0)
 
-    def test_gen_annulus_simple(self):
-        """Test generating simple annuli."""
-        pass
+    def test_bool_and_eq(self):
+        """Test __bool__ and __eq__ methods."""
+        # Test empty database
+        empty_df = pd.DataFrame({"lat": [], "lon": [], "radius": []})
+        empty_cdb = CraterDatabase(empty_df, "moon")
+        self.assertFalse(bool(empty_cdb))
 
-    def test_gen_annulus_precise(self):
-        """Test generating precise annuli."""
-        pass
+        # Test non-empty database
+        self.assertTrue(bool(self.moon_cdb))
 
-    def test_antimeridian_splitting(self):
-        """Test generating annuli when split over the antimeridian."""
-        pass
+        # Test equality with same data
+        cdb1 = self.moon_cdb.copy()
+        cdb2 = self.moon_cdb.copy()
+        self.assertEqual(cdb1, cdb2)
 
-    def test_pole_crossing(self):
-        """Test annuli that cross the North or South pole."""
-        pass
+        # Test inequality with different data
+        cdb2.data.loc[0, "lat"] = 0.0
+        self.assertNotEqual(cdb1, cdb2)
+
+        # Test inequality with different type
+        self.assertNotEqual(cdb1, "not a CraterDatabase")
+
+    def test_gen_point_edge_cases(self):
+        """Test _gen_point with various edge cases."""
+        # Test with points at poles
+        df = pd.DataFrame(
+            {"lat": [90.0, -90.0], "lon": [0.0, 180.0], "radius": [1.0, 1.0]}
+        )
+        cdb = CraterDatabase(df, "moon")
+        points = cdb._gen_point()
+        self.assertEqual(len(points), 2)
[…]
