+++ .github/workflows/test.yml
@@ -29,7 +29,7 @@ jobs:
         run: uv run pytest --cov-report=xml
   
       - name: Upload coverage to Codecov
-        if: ${{ matrix.python-version }} == '3.12' && ${{ matrix.os }} == 'ubuntu-latest'
+        if: ${{ matrix.python-version }} == '3.13' && ${{ matrix.os }} == 'ubuntu-latest'
         uses: codecov/codecov-action@v4
         env:
           CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
