+++ .github/workflows/pypi_publish.yml
@@ -20,7 +20,7 @@ jobs:
           cache-dependency-glob: uv.lock
 
       - name: Set up Python
-        run: uv python install 3.12
+        run: uv python install 3.13
 
       - name: Build
         run: uv build
