### 2025-08-21 766a802
add crater center detection scripts
