+++ craterpy/detection_scripts/detect_centers_by_thresholding.py
@@ -1,0 +1,622 @@
+#!/usr/bin/env python3
+"""
+Crater Center Correction CLI
+
+This script extracts crater tiles from a DEM, detects the actual crater centers,
+and corrects the lat/lon coordinates based on the pixel displacement.
+
+Example usage:
+
+poetry run python path/to/detect_centers_by_thresholding.py \
+    --mode extract \
+    --raster path/to/Lunar_LRO_LOLAKaguya_DEMmerge_60N60S_512ppd.vrt \
+    --database path/to/database.geojson \
+    --output-folder path/to/output/folder \
+    --roi-radius 2 \
+    --contrast-gamma 1.5 \
+    --prefer-central \
+    --num-craters 200 \
+    --threshold 0.8
+
+"""
+
+import os
+import sys
+import argparse
+import glob
+from pathlib import Path
+import pandas as pd
+import geopandas as gpd
+import numpy as np
+import cv2
+import matplotlib.pyplot as plt
+from PIL import Image
+import rasterio
+import rasterio.errors
+import rasterio.windows
+import rasterio.transform
+import rasterio.warp
+from pyproj import Transformer
+from pyproj.crs import ProjectedCRS
+from pyproj.crs.coordinate_operation import AzimuthalEquidistantConversion
+from shapely.ops import transform
+from skimage.feature import peak_local_max
+from scipy import ndimage as ndi
+from skimage.segmentation import watershed
+from skimage.measure import regionprops
+from craterpy import CraterDatabase
+
+
+def extract_tiles_with_transforms(cdb, fraster, region="", index=None, band=1, shape=(None, None), partial_tiles=False):
+    """
+    Extract projected tiles from raster with coordinate transformation info.
+    
+    Returns tiles along with the transformation information needed to convert
+    pixel coordinates back to lat/lon.
+    """
+    if not region:
+        region = "_crater_rims"
+        if region not in cdb.data.columns:
+            cdb.data[region] = cdb._gen_annulus(0, 1, precise=True)
+    
+    gdf = cdb.data
+    if isinstance(index, pd.Index):
+        gdf = gdf.loc[index]
+    elif isinstance(index, tuple):
+        gdf = gdf.iloc[index[0]:index[1]]
+    elif isinstance(index, int):
+        gdf = gdf.iloc[index:index+1]
+
+    tiles = []
+    transform_info = []
+    
+    for i, row in gdf.iterrows():
+        center = row['_center']
+        roi = row[region]
+        
+        # Create local azimuthal equidistant projection
+        local_crs = ProjectedCRS(
+            name=f"AzimuthalEquidistant({center.y:.2f}N, {center.x:.2f}E)",
+            conversion=AzimuthalEquidistantConversion(center.y, center.x),
+            geodetic_crs=cdb._crs,
+        )
+        
+        with rasterio.open(fraster) as src:
+            src_crs = src.crs
+            src_transform = src.transform
+        
+        # Define output grid
+        to_projected = Transformer.from_crs(cdb._crs, local_crs, always_xy=True).transform
+        dst_bounds = transform(to_projected, roi).bounds
+
+        dst_width, dst_height = shape
+        
+        # Determine output shape if not specified
+        if dst_height is None or dst_width is None:
+            src_res_x = src_transform.a
+            src_res_y = abs(src_transform.e)
+            one_pixel_lon = [center.x - src_res_x / 2, center.x + src_res_x / 2]
+            one_pixel_lat = [center.y - src_res_y / 2, center.y + src_res_y / 2]
+            proj_x, proj_y = to_projected(one_pixel_lon, one_pixel_lat)
+            
+            avg_res_meters = (abs(proj_x[1] - proj_x[0]) + abs(proj_y[1] - proj_y[0])) / 2.0
+
+            left, bottom, right, top = dst_bounds
[…]
