+++ craterpy/detection_scripts/detect_centers_using_slope_map.py
@@ -1,0 +1,1301 @@
+#!/usr/bin/env python3
+"""
+Dual-Circle Crater Center Detection via Slope Minimization
+
+This script extracts crater tiles, computes slope maps at native resolution, 
+and finds optimal crater positions using a dual-circle approach:
+1. Large circle (rim detection) - seeks low-slope rim areas
+2. Small concentric circle (center detection) - seeks low-slope crater floor
+
+Gradients computed at native DEM resolution before resizing,
+combined with dual-circle evaluation for precise crater center location.
+
+Example usage:
+
+poetry run python path/to/detect_centers_using_slope_map.py \
+    --raster data/Lunar_LRO_LOLAKaguya_DEMmerge_60N60S_512ppd.vrt \
+    --database data/database.geojson \
+    --output-folder path/to/output \
+    --tile-size 512 
+    --rim-thickness 15 \
+    --iterative-smooth-rounds 1 \
+    --search-radius 80 \
+    --step-size 2 \
+
+"""
+
+import os
+import sys
+import argparse
+import numpy as np
+import cv2
+import matplotlib.pyplot as plt
+from PIL import Image
+import rasterio
+import rasterio.warp
+import rasterio.windows
+import rasterio.transform
+import pandas as pd
+import geopandas as gpd
+from pyproj import Transformer
+from pyproj.crs import ProjectedCRS
+from pyproj.crs.coordinate_operation import AzimuthalEquidistantConversion
+from shapely.ops import transform
+from craterpy import CraterDatabase
+
+
+def compute_extended_finite_differences(dem, kernel_size):
+    """Compute gradients using extended finite difference stencils."""
+    half_size = kernel_size // 2
+    h, w = dem.shape
+    
+    grad_x = np.zeros_like(dem)
+    grad_y = np.zeros_like(dem)
+    
+    if kernel_size == 3:
+        # 3-point central difference
+        for i in range(1, h-1):
+            for j in range(1, w-1):
+                grad_x[i, j] = (dem[i, j+1] - dem[i, j-1]) / 2.0
+                grad_y[i, j] = (dem[i+1, j] - dem[i-1, j]) / 2.0
+                
+    elif kernel_size == 5:
+        # 5-point central difference
+        for i in range(2, h-2):
+            for j in range(2, w-2):
+                grad_x[i, j] = (-dem[i, j+2] + 8*dem[i, j+1] - 8*dem[i, j-1] + dem[i, j-2]) / 12.0
+                grad_y[i, j] = (-dem[i+2, j] + 8*dem[i+1, j] - 8*dem[i-1, j] + dem[i-2, j]) / 12.0
+                
+    elif kernel_size == 7:
+        # 7-point central difference
+        for i in range(3, h-3):
+            for j in range(3, w-3):
+                grad_x[i, j] = (dem[i, j-3] - 9*dem[i, j-2] + 45*dem[i, j-1] - 45*dem[i, j+1] + 9*dem[i, j+2] - dem[i, j+3]) / 60.0
+                grad_y[i, j] = (dem[i-3, j] - 9*dem[i-2, j] + 45*dem[i-1, j] - 45*dem[i+1, j] + 9*dem[i+2, j] - dem[i+3, j]) / 60.0
+                
+    elif kernel_size >= 9:
+        # For larger kernels, use numpy gradient with repeated application for smoothing effect
+        dem_smooth = dem.copy()
+        for _ in range(kernel_size // 3):
+            grad_y_temp, grad_x_temp = np.gradient(dem_smooth.astype(float))
[…]
