### 2025-08-21 e94f982
Add ML dependencies
