+++ pyproject.toml
@@ -18,6 +18,12 @@ dependencies = [
     "geopandas<2.0.0,>=1.0.1",
     "scipy<2.0.0,>=1.13.1",
     "antimeridian<1.0.0,>=0.4.0",
+    "tqdm>=4.67.1",
+    "torch>=2.8.0",
+    "segmentation-models-pytorch>=0.5.0",
+    "albumentations>=2.0.8",
+    "opencv-python>=4.12.0.88",
+    "scikit-image>=0.25.2",
 ]
 classifiers = [
     "Development Status :: 4 - Beta",
