### 2025-08-31 6a5f2a6
Fix docs links in README (closes #92)
