### 2025-08-31 #8663 @cjtu
Thanks for this great feedback and user testing @marcsit . I should have time to take a look and address these later this week.
