### 2025-09-01 bbcb25e
Fix plot scaling (closes #94)
