### 2025-09-01 483d075
Fix geojson import (closes #96)
