+++ craterpy/tests/test_classes.py
@@ -19,7 +19,7 @@ from pyproj import CRS
 from shapely.geometry import Point
 
 import craterpy
-from craterpy.classes import CRS_DICT, CraterDatabase
+from craterpy.classes import BODIES, CraterDatabase
 
 
 class TestCraterDatabase(unittest.TestCase):
@@ -32,6 +32,8 @@ class TestCraterDatabase(unittest.TestCase):
         self.vesta_tif = craterpy.sample_data["vesta.tif"]
         self.vesta_craters = craterpy.sample_data["vesta_craters.csv"]
         self.moon_cdb = CraterDatabase(self.moon_craters, "moon", units="km")
+        self.moon_cdb_rim = self.moon_cdb.copy()
+        self.moon_cdb_rim.add_annuli("rim", 1, 1.1)
         self.vesta_cdb = CraterDatabase(self.vesta_craters, "vesta_claudia_dp")
 
     def test_add_circles_annuli(self):
@@ -51,21 +53,27 @@ class TestCraterDatabase(unittest.TestCase):
 
     def test_get_stats(self):
         """Test getting statistics on a region for a raster."""
-        cdb = self.moon_cdb.copy()
-        cdb.add_annuli("rim", 1, 1.1)
-        stats = cdb.get_stats(self.moon_tif, "rim")
+        stats = self.moon_cdb_rim.get_stats(self.moon_tif, "rim")
         self.assertIn("count_rim", stats.columns)
         self.assertIn("Lat", stats.columns)
         self.assertFalse(any(stats["mean_rim"].isna()))  # No null values
 
-    def test_get_stats_parallel(self):
+    def test_get_stats_list_rasters(self):
+        """Test parallellization of get_stats for multiple rasters in list."""
+        stats = self.moon_cdb_rim.get_stats(
+            [self.moon_tif, self.moon_dem], ["rim"], nodata=0
+        )
+        self.assertIn("count_raster0_rim", stats.columns)
+        self.assertIn("count_raster1_rim", stats.columns)
+        self.assertFalse(any(stats["count_raster0_rim"].isna()))
+
+    def test_get_stats_multi_rasters_regions(self):
         """Test parallellization of get_stats for multiple rasters/regions."""
-        cdb = self.moon_cdb.copy()
-        cdb.add_annuli("rim", 1, 1.1)
-        stats = cdb.get_stats(
+        stats = self.moon_cdb_rim.get_stats(
             {"moon": self.moon_tif, "dem": self.moon_dem},
-            "rim",
+            ["rim"],
             ["median", "count"],
+            nodata={"moon": 0, "dem": 0},
         )
         self.assertIn("count_moon_rim", stats.columns)
         self.assertIn("count_dem_rim", stats.columns)
@@ -73,7 +81,7 @@ class TestCraterDatabase(unittest.TestCase):
 
     def test_body_crs_all(self):
         """Test that every defined CRS loads."""
-        for body in CRS_DICT:
+        for body in BODIES:
             with warnings.catch_warnings():
                 warnings.filterwarnings("ignore", message="Vesta*")
                 cdb = CraterDatabase(self.moon_craters, body)
@@ -172,7 +180,7 @@ class TestCraterDatabase(unittest.TestCase):
             self.assertIn("import_test", newdb.data.columns)
             self.assertIn("_center", newdb.data.columns)
             self.assertNotIn("geometry", newdb.data.columns)
-            self.assertEqual(newdb.body, "Moon")
+            self.assertEqual(newdb.body, "moon")
 
     def test_to_geojson_with_custom_geometry(self):
[…]
