+++ craterpy/__init__.py
@@ -2,19 +2,11 @@
 
 from importlib import metadata, resources
 
-from craterpy.classes import CRS_DICT, CraterDatabase
+from craterpy.classes import BODIES, CraterDatabase
 
-sample_data = {
-    k: str(resources.files(__name__).joinpath("data", k))
-    for k in (
-        "moon.tif",
-        "moon_craters.csv",
-        "vesta.tif",
-        "vesta_craters.csv",
-        "moon_dem.tif",
-    )
-}
-all_bodies = list(CRS_DICT.keys())
+sample_data = {k.name: k for k in resources.files(__name__).joinpath("data").iterdir()}
+all_bodies = list(BODIES.keys())
 __all__ = ["CraterDatabase", "sample_data"]
+
 __version__ = metadata.version(__package__)
 del metadata
