+++ craterpy/data/ganymede_craters.txt
@@ -1,0 +1,242 @@
+				-	-	Diameter (km)	-	-	-	-									
+Name	Crater Type	Latitude	Longitude	dome/depression	central pit/ring	rim	rim (predicted)	pedestal	ejecta	rings	Terrace  Width	Rimwall Width	Rim Height	Depth	Depth Method	Terrain	Image	Resolution	Note
+
+Gilgamesh	MR	-62	124	144	306	588	N/A	769	1073	~1000				1 to 3	"sh, st"	b	20639.15	0.6	"inner massif ring & smooth area, outer furrow ring; see 20638.14; Galileo NIMS target"
+
+Galileo-Marius Regio	MV	-21	179				<3000			>8300						dc	numerous	0.7	"multiple furrows, center obscured; see 20631.01-11; Galileo target"
+Nicholson Regio	MV	38	32				<3000			>6000						dc	numerous	1.2	"multiple furrows, center obscured; Galileo target"
+Perrine Regio	MV	56	46				<1000			~1580						dc	16405.46	1.1	"multiple furrows, center obscured; see also 16402.02"
+E. Marius Regio	MV	-28	153				<500			~500						dc	20637.50	0.8	"several furrows, smooth central feature; Galileo target"
+
+ZS23:018	PA	-23	18				301	~460								dc	16404.44	1.3	"irregular shape, crosscut by bright terrain, high-sun view"
+Thebes Facula	PA	7	201				245	~380								dc	20635.45	1.3	"irregular shape, high-sun view; Galileo target"
+Busiris Facula	PA	15	215				231	~360								dc	20635.45	1.3	"crosscut by bright terrain, high-sun view"
+Memphis Facula	PA	16	133	~61	~114	215	227	355								dc	20638.29	0.7	type example; Galileo target
+Coptos Facula	PA	10	209				207	325								dc	20635.45	1.3	"crosscut by bright terrain, high-sun view"
+ZN28:189	PA	28	189				200	315								dc	20635.37	1.3	crosscut by bright terrain
+ZS10:211	PA	-10	211				200	~315								dc	20635.56	1.3	"disrupted and crosscut by bright terrain, high-sun view"
+ZN73:138	PA	73	138				172	~275								dc	20608.11	9.6	"oblique view, low resolution"
+Siwah Facula	PA	8	143				160	257								dc	20637.11	1.0	"irregular shape, crosscut by bright terrain"
+ZN22:195	PA	22	195				155	250								dc	20635.37	1.3	crosscut by bright terrain
+ZN10:008	PA	10	8				149	240								dc	16402.22	2.1	crosscut by bright terrain
+ZN29:356	PA	29	356		49		122	~200								dc	16405.28	1.1	half buried by bright terrain
+ZS20:325	PA	-20	325	38			121	198								dc	16405.04	1.2	concentric rings; possible penepalimpsest
[…]
