### 2025-09-02 5b17f56
Refactor body crs. Add more example body images and crater lists
