### 2025-09-02 f3e2664
Parallelize with joblib and report timing (#56, #93).\nMinor bug fix where diameter sometimes treated as radius.
