+++ craterpy/classes.py
@@ -3,8 +3,6 @@ from copy import deepcopy
 from typing import Union
 import warnings
 from pathlib import Path
-import multiprocessing
-import antimeridian
 import numpy as np
 import pandas as pd
 import geopandas as gpd
@@ -13,18 +11,12 @@ import matplotlib.pyplot as plt
 import rasterio as rio
 import shapely
 from shapely.geometry import Point
-from pyproj import CRS, Transformer
-from pyproj.crs import ProjectedCRS
-from pyproj.crs.coordinate_operation import AzimuthalEquidistantConversion
-from shapely.ops import transform
-from rasterstats import zonal_stats, gen_zonal_stats
+from pyproj import CRS
+from rasterstats import gen_zonal_stats
 import cartopy.crs as ccrs
 from cartopy.feature import ShapelyFeature
 from cartopy.mpl.gridliner import LONGITUDE_FORMATTER, LATITUDE_FORMATTER
 import craterpy.helper as ch
-
-# Suppress antimeridian warning
-warnings.filterwarnings("ignore", category=antimeridian.FixWindingWarning)
 
 # Default stats for rasterstats
 STATS = ("mean", "std", "count")
@@ -173,7 +165,9 @@ class CraterDatabase:
         self.data[self._loncol] = pd.to_numeric(self.data[self._loncol])
         # Look for radius / diam column, store as _radius_m
         rcol = ch.find_rad_or_diam_col(self.data)
-        div = 2 if rcol.startswith("d") else 1  # diam -> rad conversion
+        div = (
+            2 if rcol.lower().startswith("d") else 1
+        )  # diam -> rad conversion
         self.data["_radius_m"] = pd.to_numeric(self.data[rcol]) / div
         self._radcol = "_radius_m"
 
@@ -201,6 +195,10 @@ class CraterDatabase:
             [p for p in self._get_properties() if not p.startswith("_")]
         )
         return f"CraterDatabase of length {len(self.data)} with attributes {attrs}."
+
+    def __str__(self):
+        body = self.body.capitalize()
+        return f"{body} CraterDatabase (N={len(self.data)})"
 
     def __eq__(self, other):
         """
@@ -261,106 +259,14 @@ class CraterDatabase:
         # Note: list comprehension is faster than df.apply
         return [Point(xy) for xy in zip(self.lon, self.lat)]
 
-    def _gen_annulus(self, inner, outer, precise=False, **kwargs):
+    def _gen_annulus(self, inner, outer, **kwargs):
         """Return annular geometry for each row."""
         # Don't know why it prefers to edit a copy, but _precise was giving None geometries with GeoSeries([polys_out])
         out = self.center.copy()
-        if precise:
-            out.loc[:] = self._gen_annulus_precise(
-                self.center, self.rad, inner, outer, **kwargs
-            )
-        else:
-            # Compute simple circles for all craters (fast)
-            out.loc[:] = self._gen_annulus_simple(
-                inner, outer, self._crs, self._crs180, **kwargs
-            )
-            # Be more precise for craters that are warped or go out of bounds
-            # find craters warped more than 10% in plate caree. Horizonal stretch
-            # goes as cos and cos(25)=0.906, so circles within 25N-25S are <10% err
-            equatorial = (-25 < out.bounds.miny) & (out.bounds.maxy < 25)
-            oob = out.bounds.maxx - out.bounds.minx >= 300
-            out.loc[~equatorial | oob] = self._gen_annulus_precise(
-                self.center[~equatorial | oob],
-                self.rad[~equatorial | oob],
-                inner,
-                outer,
-            )
-        return out
-
-    def _get_annular_buffer(self, pt, rad, inner, outer, nvert=32):
[…]
