### 2025-09-02 af22bcf
Parallelize with joblib and report timing (#56, #93).\nMinor bug fix where diameter sometimes treated as radius.
