+++ craterpy/helper.py
@@ -1,8 +1,24 @@
 """This file contains various helper functions for craterpy"""
 
+import contextlib
+import warnings
 import numpy as np
 import pandas as pd
 from scipy.spatial import cKDTree
+import joblib
+from tqdm import tqdm
+import antimeridian
+from rasterstats import zonal_stats
+from joblib import Parallel, delayed
+from shapely.geometry import Point
+from shapely.ops import transform
+from pyproj import Transformer
+from pyproj.crs import ProjectedCRS
+from pyproj.crs.coordinate_operation import AzimuthalEquidistantConversion
+
+# Suppress antimeridian warning
+warnings.filterwarnings("ignore", category=antimeridian.FixWindingWarning)
+warnings.filterwarnings("ignore", "Setting nodata.*", module=r".*rasterstats")
 
 
 # Geospatial helpers
@@ -203,6 +219,210 @@ def inglobal(lat, lon):
 
     """
     return (-90 <= lat <= 90) and (0 <= lon360(lon) <= 360)
+
+
+# Shape geometry helpers
+def gen_annuli(centers, rads, inner, outer, crs, **kwargs):
+    """
+    Generate annuli using a local azimuthal equidistant projection for each crater.
+    This version processes craters in parallel to improve performance.
+    """
+    # Use joblib to run _create_single_annulus on all available CPU cores.
+    # n_jobs=-1 means use all available cores except one.
+    n_jobs = kwargs.get("n_jobs", -2)
+    tasks = (
+        delayed(create_single_annulus)(
+            center, rad, inner, outer, crs, **kwargs
+        )
+        for center, rad in zip(centers, rads)
+    )
+    with tqdm_joblib(tqdm(desc="Generating polygons", total=len(centers))):
+        annuli = Parallel(n_jobs=n_jobs)(tasks)
+    return annuli
+
+
+def create_single_annulus(center, rad, inner, outer, geodetic_crs, **kwargs):
+    """
+    Processes a single center point to create a precise geodetic annulus.
+    (This contains the logic from your original for loop)
+    """
+    # Create a local azimuthal equidistant projection for the crater
+    local_crs = ProjectedCRS(
+        name=f"AzimuthalEquidistant({center.y:.2f}N, {center.x:.2f}E)",
+        conversion=AzimuthalEquidistantConversion(center.y, center.x),
+        geodetic_crs=geodetic_crs,
+    )
+
+    # Generate the buffer in the local, flat projection
+    # Assumes self._get_annular_buffer is available or moved here
+    buf = get_annular_buffer(Point(0, 0), rad, inner, outer, **kwargs)
+
+    # Create transformer and unproject the buffer to the geodetic CRS
+    to_geodetic = Transformer.from_crs(
+        local_crs, geodetic_crs, always_xy=True
+    ).transform
+    annulus = transform(to_geodetic, buf)
+
+    # Fix antimeridian wrapping if necessary
+    if annulus.bounds[2] - annulus.bounds[0] >= 300:
+        annulus = fix_antimeridian_wrap(
+            annulus, rad, outer, geodetic_crs, local_crs, **kwargs
+        )
+
+    return annulus
+
+
+def get_annular_buffer(point, rad, inner, outer, **kwargs):
+    """ """
+    # Num vertices to make each circular poly
+    nvert = kwargs.get("nvert", 32)
+    outer_scaled = rad * outer
+    inner_scaled = rad * inner
+    outer_poly = point.buffer(outer_scaled, quad_segs=nvert // 4)
+    if inner_scaled <= 0:
+        return outer_poly
+    inner_poly = point.buffer(inner_scaled, quad_segs=nvert // 4)
[…]
