### 2025-09-03 e3c5ce5
Replace black linter with ruff and precommit
