+++ CONTRIBUTING.rst
@@ -4,36 +4,23 @@ Thank you for your interest in contributing to craterpy!
 
 Who can contribute?
 -------------------
-*Anyone!* This project is as much a tool to learn about open source science as it is a tool to improve planetary data science workflows. If you've never contributed to open source before, great! Drop us a line! If you're an expert hacker and think we're doing everything wrong, awesome! We want to hear from you too!
+*You!* Bug reports, user feedback, and feature requests are very real contributions and begin over on the `issues`_ board.
 
-Why contribute?
----------------
-If you're interested in any of the following:
-
-- planets
-- science
-- open source
-- contributing to active research
-- learning in a supportive environment
-
-Or you already use craterpy and want to make it better for everyone!
+To contribute to the code directly, even if it's your first time contributing to open source, see below.
 
 Where to start?
 ---------------
-A good place to start is by scanning the `issue tracker`_ and seeing if there is something that you are interested in taking on. Drop a comment on an issue to let us know that you want to work on it, or leave a question if something needs to be clarified.
+A good place to start is by scanning the open `issues`_ and seeing if there is something that you are interested in taking on. Drop a comment on an issue to let us know that you want to work on it, or leave a question if something needs to be clarified.
 
-.. _`issue tracker`: https://github.com/cjtu/craterpy/issues
+.. _`issues`: https://github.com/cjtu/craterpy/issues
 
 How to get started?
 -------------------
-If you made it this far and you're still interested, that's great! Here are four simple steps to get started contributing to craterpy:
+You will need to download the repo and install the development environment locally to start contributing code.
 
-1. Fork the `main repository`_ by clicking the link and clicking the Fork button in the upper right corner (read about `forking`_).
+1. Fork the `main repository`_ by clicking the link and clicking the Fork button in the upper right corner.::
 
-2. Clone the forked repository to your machine (Git must be `installed`_). ::
-
-    git clone https://github.com/your_username/craterpy.git
-    cd craterpy
+2. Clone the forked repository to your machine. Check out `GitHub Desktop`_ if it's your first time using Git. ::
 
 3. Install uv from https://docs.astral.sh/uv/#installation then download dependencies using uv::
 
@@ -54,59 +41,33 @@ Preparing a commit
 
 When you are ready to contribute your changes, make sure to:
 
-1. Add unittests to the tests folder.
+1. Add unittests to the tests folder.::
 
 2. Test your changes::
 
     uv run pytest craterpy
 
-3. Apply automatic code formatting with black::
-
-    uv run black craterpy
-
-4. Add and commit your changes::
+3. Add and commit your changes::
     
     git add . 
     git commit -m "Descriptive commit message here" 
 
-Now you're ready to push your changes and open a Pull Request on GitHub!
+4. The `ruff` pre-commit hook may reformat some of your files to ensure a standard python style. If changes are made, repeat step 3 to add and commit your changes.::
+
+5. Push your branch and open a Pull Request on GitHub!
 
 .. _`main repository`: https://github.com/cjtu/craterpy
-.. _`forking`: https://guides.github.com/activities/forking/
-.. _`installed`: https://git-scm.com/downloads
-.. _`Anaconda`: https://www.anaconda.com/download/
+.. _`Github Desktop`: https://github.com/apps/desktop
+.. _`GitHub Workflow`: https://docs.github.com/en/get-started/using-github/github-flow
 
 
-Submitting a Pull Request
--------------------------
-Three simple steps:
+Thanks fro contributing!
+------------------------
+Sit tight while a collaborator reviews your contribution. We might ask you to clarify some code or add tests and will work with you to get your contribution merged!
 
[…]
