### 2025-09-03 4ddc60f
Replace black linter with ruff and precommit
