+++ craterpy/helper.py
@@ -2,19 +2,20 @@
 
 import contextlib
 import warnings
+
+import antimeridian
+import joblib
 import numpy as np
 import pandas as pd
-from scipy.spatial import cKDTree
-import joblib
-from tqdm import tqdm
-import antimeridian
-from rasterstats import zonal_stats
 from joblib import Parallel, delayed
-from shapely.geometry import Point
-from shapely.ops import transform
 from pyproj import Transformer
 from pyproj.crs import ProjectedCRS
 from pyproj.crs.coordinate_operation import AzimuthalEquidistantConversion
+from rasterstats import zonal_stats
+from scipy.spatial import cKDTree
+from shapely.geometry import MultiPolygon, Point, Polygon
+from shapely.ops import transform
+from tqdm import tqdm
 
 # Suppress antimeridian warning
 warnings.filterwarnings("ignore", category=antimeridian.FixWindingWarning)
@@ -180,10 +181,7 @@ def greatcircdist(lat1, lon1, lat2, lon2, radius):
     lat1, lon1, lat2, lon2 = np.radians([lat1, lon1, lat2, lon2])
     # Haversine
     dlat, dlon = abs(lat2 - lat1), abs(lon2 - lon1)
-    a = (
-        np.sin(dlat / 2) ** 2
-        + np.cos(lat1) * np.cos(lat2) * np.sin(dlon / 2) ** 2
-    )
+    a = np.sin(dlat / 2) ** 2 + np.cos(lat1) * np.cos(lat2) * np.sin(dlon / 2) ** 2
     theta = 2 * np.arcsin(np.sqrt(a))
     return radius * theta
 
@@ -222,29 +220,73 @@ def inglobal(lat, lon):
 
 
 # Shape geometry helpers
-def gen_annuli(centers, rads, inner, outer, crs, **kwargs):
+def gen_annuli(centers: list, rads: list, inner: float, outer: float, crs, **kwargs):
     """
-    Generate annuli using a local azimuthal equidistant projection for each crater.
-    This version processes craters in parallel to improve performance.
+    Generate annular polygons for a set of crater centers using a local azimuthal equidistant projection.
+
+    Each annulus is generated by projecting to a local coordinate system, buffering, and then
+    transforming back to geodetic coordinates.
+
+    Processes in parallel with joblib.
+
+    Parameters
+    ----------
+    centers : list of shapely.geometry.Point
+        List of crater center points (latitude/longitude).
+    rads : list of float
+        List of crater radii (in projection units).
+    inner : float
+        Inner radius scaling factor (relative to crater radius).
+    outer : float
+        Outer radius scaling factor (relative to crater radius).
+    crs : pyproj.CRS
+        Geodetic coordinate reference system for output polygons.
+    **kwargs
+        Additional keyword arguments passed to annulus creation (e.g., n_jobs, nvert).
+
+    Returns
+    -------
+    list of shapely.geometry.Polygon
+        List of annular polygons in geodetic coordinates.
     """
-    # Use joblib to run _create_single_annulus on all available CPU cores.
-    # n_jobs=-1 means use all available cores except one.
+    # Use joblib to run _create_single_annulus on all available CPU cores minus 1.
     n_jobs = kwargs.get("n_jobs", -2)
     tasks = (
-        delayed(create_single_annulus)(
-            center, rad, inner, outer, crs, **kwargs
-        )
-        for center, rad in zip(centers, rads)
+        delayed(create_single_annulus)(center, rad, inner, outer, crs, **kwargs)
+        for center, rad in zip(centers, rads, strict=True)
     )
     with tqdm_joblib(tqdm(desc="Generating polygons", total=len(centers))):
[…]
