+++ craterpy/__init__.py
@@ -1,7 +1,8 @@
 """craterpy module"""
 
 from importlib import metadata, resources
-from craterpy.classes import CraterDatabase, CRS_DICT
+
+from craterpy.classes import CRS_DICT, CraterDatabase
 
 sample_data = {
     k: str(resources.files(__name__).joinpath("data", k))
@@ -14,6 +15,6 @@ sample_data = {
     )
 }
 all_bodies = list(CRS_DICT.keys())
-
+__all__ = ["CraterDatabase", "sample_data"]
 __version__ = metadata.version(__package__)
 del metadata
