+++ craterpy/classes.py
@@ -1,21 +1,23 @@
 import json
-from copy import deepcopy
-from typing import Union
 import warnings
+from copy import deepcopy
 from pathlib import Path
-import numpy as np
-import pandas as pd
+from typing import Union
+
+import cartopy.crs as ccrs
 import geopandas as gpd
 import matplotlib.image as mpli
 import matplotlib.pyplot as plt
+import numpy as np
+import pandas as pd
 import rasterio as rio
 import shapely
-from shapely.geometry import Point
+from cartopy.feature import ShapelyFeature
+from cartopy.mpl.gridliner import LATITUDE_FORMATTER, LONGITUDE_FORMATTER
 from pyproj import CRS
 from rasterstats import gen_zonal_stats
-import cartopy.crs as ccrs
-from cartopy.feature import ShapelyFeature
-from cartopy.mpl.gridliner import LONGITUDE_FORMATTER, LATITUDE_FORMATTER
+from shapely.geometry import Point
+
 import craterpy.helper as ch
 
 # Default stats for rasterstats
@@ -165,9 +167,7 @@ class CraterDatabase:
         self.data[self._loncol] = pd.to_numeric(self.data[self._loncol])
         # Look for radius / diam column, store as _radius_m
         rcol = ch.find_rad_or_diam_col(self.data)
-        div = (
-            2 if rcol.lower().startswith("d") else 1
-        )  # diam -> rad conversion
+        div = 2 if rcol.lower().startswith("d") else 1  # diam -> rad conversion
         self.data["_radius_m"] = pd.to_numeric(self.data[rcol]) / div
         self._radcol = "_radius_m"
 
@@ -184,16 +184,12 @@ class CraterDatabase:
 
         # Set geometry and covert to GeoDataFrame if not already
         if not isinstance(self.data, gpd.GeoDataFrame):
-            self.data = gpd.GeoDataFrame(
-                self.data, geometry="_center", crs=self._crs
-            )
+            self.data = gpd.GeoDataFrame(self.data, geometry="_center", crs=self._crs)
         elif "geometry" not in self.data.columns:
             self.data.set_geometry("_center", inplace=True)
 
     def __repr__(self):
-        attrs = ", ".join(
-            [p for p in self._get_properties() if not p.startswith("_")]
-        )
+        attrs = ", ".join([p for p in self._get_properties() if not p.startswith("_")])
         return f"CraterDatabase of length {len(self.data)} with attributes {attrs}."
 
     def __str__(self):
@@ -257,7 +253,7 @@ class CraterDatabase:
     def _gen_point(self):
         """Return point geometry (lon, lat) for each row."""
         # Note: list comprehension is faster than df.apply
-        return [Point(xy) for xy in zip(self.lon, self.lat)]
+        return [Point(xy) for xy in zip(self.lon, self.lat, strict=True)]
 
     def _gen_annulus(self, inner, outer, **kwargs):
         """Return annular geometry for each row."""
@@ -286,11 +282,11 @@ class CraterDatabase:
         body = body.lower()
         if body in ["vesta_claudia_dp", "vesta_claudia_double_prime"]:
             return "vesta_claudia_dp", 0
-        elif body in ["vesta_claudia_p", "vesta_claudia_prime"]:
+        if body in ["vesta_claudia_p", "vesta_claudia_prime"]:
             return "vesta_claudia_p", 190
-        elif body in ["vesta_claudia", "vesta_dawn_claudia"]:
+        if body in ["vesta_claudia", "vesta_dawn_claudia"]:
             return "vesta_claudia", 150
-        elif body in ["vesta_iau_2000", "vesta_iau2000"]:
+        if body in ["vesta_iau_2000", "vesta_iau2000"]:
[…]
