+++ .github/workflows/code_quality_checks.yml
@@ -1,4 +1,3 @@
----
 name: Code Quality Checks
 
 on:
@@ -22,7 +21,5 @@ jobs:
       - name: Install the project
         run: uv sync --locked --all-extras --dev
 
-      - name: Check format with black
-        run: uv run black --check craterpy/
-
-        
+      - name: Check format with Ruff
+        run: uv run ruff check
