+++ .pre-commit-config.yaml
@@ -1,0 +1,23 @@
+repos:
+- repo: local
+  hooks:
+  - id: uv-lock
+    name: Update uv lock file
+    entry: uv lock
+    language: system
+    pass_filenames: false
+  - id: ruff-fix
+    name: Ruff check and fix
+    entry: uv run ruff check
+    args: [--fix]
+    language: system
+    types: [python]
+  - id: ruff-format
+    name: Ruff format
+    entry: poetry run ruff format
+    language: system
+    types: [python]
+- repo: https://github.com/pre-commit/pre-commit-hooks
+  rev: v2.3.0
+  hooks:
+  -   id: check-yaml

[…]
