+++ craterpy/classes.py
@@ -2,7 +2,6 @@ import json
 import warnings
 from copy import deepcopy
 from pathlib import Path
-from typing import Union
 
 import cartopy.crs as ccrs
 import geopandas as gpd
@@ -14,38 +13,15 @@ import rasterio as rio
 import shapely
 from cartopy.feature import ShapelyFeature
 from cartopy.mpl.gridliner import LATITUDE_FORMATTER, LONGITUDE_FORMATTER
-from pyproj import CRS
+from pyproj import CRS, Transformer
 from rasterstats import gen_zonal_stats
 from shapely.geometry import Point
 
 import craterpy.helper as ch
+from craterpy.crs import ALL_BODIES, PLANETARY_CRS, get_crs
 
 # Default stats for rasterstats
 STATS = ("mean", "std", "count")
-
-# CRS for units / coord transformations (convention is only Ocentric)
-BODIES = {
-    "mercury": 199,
-    "venus": 299,
-    "moon": 301,
-    "earth": 399,
-    "phobos": 401,
-    "mars": 499,
-    "ceres": 2000001,
-    "vesta": 2000004,
-    "europa": 502,
-    "ganymede": 503,
-    "callisto": 504,
-    "enceladus": 602,
-    "tethys": 603,
-    "dione": 604,
-    "rhea": 605,
-    "titan": 606,
-    "iapetus": 608,
-    "triton": 801,
-    "charon": 901,
-    "pluto": 999,
-}
 
 
 class CraterDatabase:
@@ -70,19 +46,12 @@ class CraterDatabase:
     # geometry columns on the fly for computing stats
 
     # Private Attrs:
-    # _crs (str): Coordinate reference system for the body.
-    # _crs180 (str): CRS with longitude in -180 to 180 degrees.
-    # _crs360 (str): CRS with longitude in 0 to 360 degrees.
-    # _crsnorth (str): CRS for the northern hemisphere.
-    # _crssouth (str): CRS for the southern hemisphere.
-    # _latcol (str): Column name for latitude.
-    # _loncol (str): Column name for longitude.
-    # _radcol (str): Column name for radius.
-    # _vesta_coord (str, optional): Coordinate system for Vesta, if applicable.
+    # _crs (str): Internal coordinate reference system for the body, defaults to planetocentric.
     def __init__(
         self,
-        dataset: Union[str, pd.DataFrame],
+        dataset: str | pd.DataFrame,
         body: str,
+        input_crs: str | CRS = "default",
         units: str = "m",
     ):
         """
@@ -91,74 +60,71 @@ class CraterDatabase:
         Parameters
         ----------
         dataset : str or pandas.DataFrame
-            if str, path to the file containing crater data.
-            if DataFrame, DataFrame containing crater data.
+            Path to the crater file or DataFrame.
         body : str
-            Planetary body, e.g. 'Moon', 'Vesta'
+            Planetary body (case-insensitive), e.g. 'Moon', 'mars', 'Vesta'
+        input_crs : str or pyproj.CRS
+            The Coordinate Reference System of the input latitude/longitude data.
+            Can be a user-friendly string alias (e.g., 'planetographic', 'claudia_dp')
+            or a valid pyproj.CRS object.
         units : str
-            Length units of radius/diameter, 'm' or 'km' (default: 'm')
+            Length units of radius/diameter column, 'm' or 'km' (default: 'm')
 
         Raises
         ------
             ValueError
                 If dataset is not a file or is not a pandas.DataFrame.
         """
-        body = body.lower()
-        lon_offset = 0  # Standardizes coordinate systems (for Vesta) if necessary
-        if "vesta" in body:
-            body, lon_offset = self._vesta_check(body)
[…]
