### 2025-09-12 9446a1c
Update crater examples and crs handling
