+++ craterpy/__init__.py
@@ -5,7 +5,8 @@ from importlib import metadata, resources
 from craterpy.classes import CraterDatabase
 from craterpy.crs import ALL_BODIES
 
-sample_data = {k.name: k for k in resources.files(__name__).joinpath("data").iterdir()}
+data_dir = resources.files(__name__).joinpath("data")
+sample_data = {p.name: p for p in data_dir.rglob("*") if p.suffix in (".csv", ".tif")}
 all_bodies = ALL_BODIES
 __all__ = ["CraterDatabase", "sample_data"]
 
