+++ craterpy/crs.py
@@ -48,45 +48,38 @@ PLANETARY_CRS = {
     },
     "enceladus": {
         "planetocentric": CRS("IAU_2015:60200"),
+        "planetographic": CRS(CRS("IAU_2015:60200").to_proj4() + " +axis=wnu"),
     },
     "tethys": {
         "planetocentric": CRS("IAU_2015:60300"),
+        "planetographic": CRS(CRS("IAU_2015:60300").to_proj4() + " +axis=wnu"),
     },
     "dione": {
         "planetocentric": CRS("IAU_2015:60400"),
+        "planetographic": CRS(CRS("IAU_2015:60400").to_proj4() + " +axis=wnu"),
     },
     "rhea": {
         "planetocentric": CRS("IAU_2015:60500"),
+        "planetographic": CRS(CRS("IAU_2015:60500").to_proj4() + " +axis=wnu"),
     },
     "iapetus": {
         "planetocentric": CRS("IAU_2015:60800"),
         "planetographic": CRS("IAU_2015:60801"),
-    },
-    "triton": {
-        "planetocentric": CRS("IAU_2015:80100"),
-    },
-    "charon": {
-        "planetocentric": CRS("IAU_2015:90100"),
     },
     "pluto": {
         "planetocentric": CRS("IAU_2015:99900"),
     },
 }
 
-# Set the default CRS fro the body, else assume planetographic
+# Set the default CRS fro the body, else assume planetocentric
 DEFAULT_CRS = {
-    "charon": "planetocentric",
-    "dione": "planetocentric",
-    "enceladus": "planetocentric",
-    "europa": "planetocentric",
-    "mars": "planetocentric",
-    "moon": "planetocentric",
-    "pluto": "planetocentric",
-    "rhea": "planetocentric",
-    "tethys": "planetocentric",
-    "triton": "planetocentric",
-    "vesta": "claudia_dp",
-    "venus": "planetocentric",
+    "ceres": "planetographic",
+    "ganymede": "planetographic",
+    "enceladus": "planetographic",
+    "tethys": "planetographic",
+    "dione": "planetographic",
+    "rhea": "planetographic",
+    "iapetus": "planetographic",
 }
 
 ALL_BODIES = list(PLANETARY_CRS.keys())
@@ -103,7 +96,7 @@ def get_crs(body: str, system: str | CRS = "default") -> CRS:
     except CRSError:
         try:
             system = (
-                DEFAULT_CRS.get(body, "planetographic")
+                DEFAULT_CRS.get(body, "planetocentric")
                 if system == "default"
                 else str(system)
             )
