### 2025-09-12 4145ebb
Update crater examples and crs handling
