+++ craterpy/data/callisto_craters.txt
@@ -1,139 +1,0 @@
-				-	-	Diameter (km)	-	-	-										
-Name	Crater Type	Latitude	Longitude	central dome	central pit/ring	rim	rim (predicted)	pedestal	ejecta	rings	Terrace Width	Rimwall Width	Rim Height	Depth	Depth Method	Terrain	Image	Resolution	Note
-
-Valhalla	MV	15	56				980		~1800	~3800						dc	numerous	1.0	"numerous furrows, ridges, central bright floor deposit; Galileo target"
-Asgard	MV	32	142				660		~1400	~1880						dc	20606.21	3.0	"numerous furrows, central bright floor deposit; Galileo target; see also 16428.09"
-ZS05:232	MV	-5	232							~550						dc	20617.21	1.7	furrows
-ZN34:358	MV	34	358							~350						dc	16421.43	1.7	furrows; Galileo target
-
-ZS64:350	PA	-64	350				295	~450								dc	16418.06	2.7	oblique view; Galileo target
-ZS26:158	PA	-26	158				227	350								dc	20606.33	3.0	concentric outer dark lineament
-ZN04:021	PA	4	21		52		137	220								dc	16421.47	1.7	"somewhat irregular shape, central bright ring or dome"
-ZN14:351	PA	14	351				123	200								dc	16421.43	1.7	"irregular shape, concentric and radial lineaments; Galileo target"
-ZN26:018	PA	26	18	68			99	165											
-
-Adlinda 'A'	PP	-47	31				410	~600								dc	16418.14	2.7	"outer scarp, concentric ridges, smooth texture, see also 16421.21; Galileo target"
-Asgard 'A'	PP	46	135				370	550								dc	16428.07	1.3	"outer scarp, ridges, smooth, superposed on Asgard, image smear; see 20606.25; Galileo target"
-
-Adlinda	UC	-56	23			330										dc	16418.06	2.7	bright floor/rays; Galileo target
-Adlinda 'B'	UC	-64	355			200										pa	16418.06	2.7	"large bright spot (image saturated), superposed on PA, oblique view; Galileo target"
-ZS02:189	UC	-2	189			160	157	~250								dc	20606.33	3.0	"irregular circular lineament, embayed by dark material"
-ZN52:274	UC	52	274			140										dc	20616.45	1.7	degraded circular ridge or scarp
-
-Buri	AD	-40	44	52	~110	250	N/A									dc	16418.14	2.7	"bright floor, dark halo craters"
-ZN54:083	AD	54	83	55	98	200	N/A									dc	16424.50	1.0	"bright floor, radial grooves"
-Grimr	AD	41	215	60	95	200	N/A									dc	20616.53	1.7	high-sun view
-ZS29:180	AD	-29	180	53		180	N/A									dc	20606.37	3.0	"bright floor, dark halo craters, high-sun view"
-Alfr	AD	-11	223		~90	180	N/A									dc	20617.21	1.7	high-sun view
[…]
