### 2025-09-15 4b375a9
Update example files and fix bug in plot_rois
