### 2025-09-15 9995902
Update example files and fix bug in plot_rois
