+++ craterpy/tests/test_classes.py
@@ -35,7 +35,7 @@ class TestCraterDatabase(unittest.TestCase):
         self.moon_cdb = CraterDatabase(self.moon_craters, "moon", units="km")
         self.moon_cdb_rim = self.moon_cdb.copy()
         self.moon_cdb_rim.add_annuli("rim", 1, 1.1)
-        self.vesta_cdb = CraterDatabase(self.vesta_craters, "vesta", "claudia_dp", "km")
+        self.vesta_cdb = CraterDatabase(self.vesta_craters, "vesta", units="km")
 
     def test_add_circles_annuli(self):
         """Test adding annular geometries to CraterDataBase."""
@@ -56,7 +56,7 @@ class TestCraterDatabase(unittest.TestCase):
         """Test getting statistics on a region for a raster."""
         stats = self.moon_cdb_rim.get_stats(self.moon_tif, "rim")
         self.assertIn("count_rim", stats.columns)
-        self.assertIn("Lat", stats.columns)
+        self.assertIn("Latitude", stats.columns)
         self.assertFalse(any(stats["mean_rim"].isna()))  # No null values
 
     def test_get_stats_list_rasters(self):
@@ -80,7 +80,7 @@ class TestCraterDatabase(unittest.TestCase):
         self.assertIn("count_dem_rim", stats.columns)
         self.assertFalse(any(stats["median_moon_rim"].isna()))
 
-    def test_body_crs_all(self):
+    def test_body_all_bodies_default_crs(self):
         """Test that every defined CRS loads."""
         for body in ALL_BODIES:
             with warnings.catch_warnings():
@@ -91,18 +91,16 @@ class TestCraterDatabase(unittest.TestCase):
 
     def test_vesta_coord_correction(self):
         """Test Vesta's various coordinate systems."""
-        df = pd.DataFrame({"lat": [0, 10], "lon": [-90, 120], "radius": [1, 2]})
-        cdb = CraterDatabase(df, "vesta_claudia_double_prime")  # 0 offset
+        df = pd.DataFrame({"lat": [0, -1.6], "lon": [-90, 146], "radius": [1, 2]})
+        cdb = CraterDatabase(df, "vesta", "claudia_dp")  # 0 offset
         self.assertEqual(cdb.lon.iloc[0], -90)
-        self.assertEqual(cdb.lon.iloc[1], 120)
-        cdb = CraterDatabase(df, "vesta_claudia")  # +150
-        self.assertEqual(cdb.lon.iloc[0], 60)
-        self.assertEqual(cdb.lon.iloc[1], -90)
-        cdb = CraterDatabase(df, "vesta_claudia_prime")  # +190
-        self.assertEqual(cdb.lon.iloc[0], 100)
-        self.assertEqual(cdb.lon.iloc[1], -50)
-        with self.assertRaises(NotImplementedError):
-            cdb = CraterDatabase(df, "vesta_iau_2000")
+        self.assertEqual(cdb.lon.iloc[1], 146)
+        cdb = CraterDatabase(df, "vesta", "claudia_p")  # -10
+        self.assertAlmostEqual(cdb.lon.iloc[0], -100)
+        self.assertAlmostEqual(cdb.lon.iloc[1], 136)
+        cdb = CraterDatabase(df, "vesta", "dawn_claudia")  # +210
+        self.assertAlmostEqual(cdb.lon.iloc[0], 120)
+        self.assertAlmostEqual(cdb.lon.iloc[1], -4)
 
     def test_units(self):
         """Test importing with radii in m or km."""
@@ -174,13 +172,11 @@ class TestCraterDatabase(unittest.TestCase):
             cdb.to_geojson(fcenter.name)
             gdf = gpd.read_file(fcenter.name)  # inspect the written file
             self.assertEqual(len(cdb), len(gdf))
[…]
