### 2025-09-15 6a89939
Add planet examples to docs
