### 2025-09-15 cba03eb
Add planet examples to docs
