+++ craterpy/data/craters/README.md
@@ -1,0 +1,5 @@
+# Sample data craters
+
+Crater lists were drawn from the [USGS Gazetteer of Planetary Nomenclature](https://planetarynames.wr.usgs.gov/) which compiles named features on planetary bodies approved by the IAU, accessed Sep 2025. These crater lists are not intended to be complete or precise and are provided only for testing and example usage of craterpy.
+
+Queries for named craters and basins on each body were perfomed and compiled, trimmed to 2 digit precision CSV files and saved with diam(km),lat,lon columns.
