+++ pyproject.toml
@@ -51,6 +51,8 @@ dev = [
     "pytest-cov<7.0.0,>=6.0.0",
     "ruff>=0.12.11",
     "pre-commit>=4.3.0",
+]
+docs = [
     "myst-nb>=1.3.0",
     "myst-parser<2.0.0,>=1.0.0",
     "sphinx<7.0.0,>=6.2.1",
