### 2025-09-17 86b5e7a
Improve docs and readme (#93)
