+++ docs/source/getting_started.md
@@ -1,91 +1,122 @@
 # Getting Started
 
-The basis of `craterpy` is the `CraterDatabase` which reads, digitizes, plots, and extracts statistics from supplied crater locations.
+The basis of `craterpy` is the `CraterDatabase` which reads crater data, manages coordinate reference systems (CRS), digitizes regions of interest, and plots or extracts statistics from supplied crater locations.
 
-```python
-from craterpy import CraterDatabase
-cdb = CraterDatabase("/path/to/craters.csv", body="Moon")
-```
+## Importing Craters
 
-## Importing a list of craters
-
-Instantiating a `CraterDatabase` requires a table of craters, the planetary body, and the length units of the radius/diameter column ("m" or "km"). Data can be supplied as a file path or `pandas.DataFrame` that must contain the following named columns in any order (case-insensitive):
-
-- `Latitude` or `Lat`
-- `Longitide` or `Lon`
-- `Radius` or `Rad` or `Diameter` or `Diam`
-
-When multiple columns match, `craterpy` uses the first column found.
+The `CraterDatabase` takes a table of craters with at least columns for latitude, longitude, and radius or diameter of the craters. The exact names of the columns are not strict, ex 'Lat' and 'rad' are recognized, but if multiple columns match, `craterpy` will assume the first is correct.
 
 ```{note}
 To choose specific columns or filter to desired rows, first read the table into a `pandas.DataFrame` and pass the filtered subset dataframe to `CraterDatabase`.
 ```
 
-Here we read a sample list of lunar craters, filter it to larger than 55 km and import it into a `CraterDatabase`:
+Here we read a sample list of lunar craters, optionally filter it to keep only the largest and then improt it into a `CraterDatabase`:
 
 ```python
-from craterpy import CraterDatabase, sample_data as sd
+from craterpy import CraterDatabase, sample_data
 import pandas as pd
-df = pd.read_csv(sd["moon_craters.csv"])
-df = df[df['Rad'] > 55]
+crater_file = sample_data["moon_craters_km.csv"]
+cdb_all = CraterDatabase(crater_file, body="Moon", units="km")
+print(cdb_all)
+# Moon CraterDatabase (N=786)
+
+# Or filter/clean data with pandas first
+df = pd.read_csv(sample_data["moon_craters_km.csv"])
+df = df[df["Diameter (km)"] > 100]
 cdb = CraterDatabase(df, body="Moon", units="km")
 print(cdb)
-# CraterDatabase of length 203 with attributes lat, lon, rad, center.
+# Moon CraterDatabase (N=222)
 ```
 
-Planetary `body` is required to ensure the correct coordinate reference system (CRS) is used. Supported bodies can be printed with:
+Planetary `body` is a required `CraterDatabase` parameter to ensure the correct coordinate reference system (CRS) is used (read more about CRS [here](https://docs.qgis.org/3.40/en/docs/gentle_gis_introduction/coordinate_reference_systems.html)). Supported bodies are given in `craterpy.all_bodies`. The CRS defaults to the `IAU_2015` CRS most commonly used for the particular body, see list [here](https://planetarynames.wr.usgs.gov/TargetCoordinates).  A custom input CRS for the crater coordinates can be specified as follows:
 
 ```python
-import craterpy
-print(craterpy.all_bodies)
-# ['moon', 'mars', 'mercury', 'venus', 'europa', 'ceres', 'vesta']
[…]
