+++ .pre-commit-config.yaml
@@ -21,3 +21,8 @@ repos:
   rev: v2.3.0
   hooks:
   -   id: check-yaml
+- repo: https://github.com/kynan/nbstripout
+  rev: 0.5.0
+  hooks:
+    - id: nbstripout
+      files: ^docs/
