### 2025-09-17 d17afb6
Improve docs and readme (#93)
