### 2025-09-24 0a12ef4
Fix sign error in vesta CRS
