+++ craterpy/crs.py
@@ -33,9 +33,9 @@ PLANETARY_CRS = {
         # Claudia double prime (same as IAU_2015)
         "claudia_dp": CRS("IAU_2015:200000400"),
         # Claudia Prime, Claudia crater at 136 E
-        "claudia_p": CRS.from_proj4("+proj=longlat +R=255000 +lon_0=-10 +no_defs"),
+        "claudia_p": CRS.from_proj4("+proj=longlat +R=255000 +lon_0=+10 +no_defs"),
         # Dawn Claudia, Claudia crater at 356 E
-        "dawn_claudia": CRS.from_proj4("+proj=longlat +R=255000 +lon_0=210 +no_defs"),
+        "dawn_claudia": CRS.from_proj4("+proj=longlat +R=255000 +lon_0=-210 +no_defs"),
         # Not Implemented: IAU-2000, Claudia crater at 4.3N, 145 E
     },
     "europa": {
