### 2025-09-26 #8663 @jGaboardi
@cjtu – Please complete the [5 tasks above](https://github.com/openjournals/joss-reviews/issues/8663#issuecomment-3336428578) in the *Additional Author Tasks* section of the *Post-Review Checklist*.

Once you have completed these, ping me back here and I will start on my 7 tasks.

We're getting close!
