+++ CONTRIBUTORS.md
@@ -4,5 +4,6 @@ Contributors in the order of first contribution:
 - Alexandre Boivin
 - Dylan Hickson
 - Cailin Gallinger
+- Cong
 - Ari Essunfeld
 
