### 2025-09-26 #8663 @cjtu
Hi @jGaboardi, I've completed the author checklist above and note:

- Fixed a discrepancy in the author list, now all authors, affiliations and OrcIDs are consistent between the paper.md and Zenodo
- Released the latest version of the package as v0.10.0 ([releases](https://github.com/cjtu/craterpy/releases/tag/0.10.0)), archived on Zenodo (DOI: https://doi.org/10.5281/zenodo.17211085). Full citation below:
- > Tai Udovicic, C. J., Essunfeld, A., & Costello, E. S. (2025). Craterpy: Impact crater data science in Python (0.10.0). Zenodo. https://doi.org/10.5281/zenodo.17211085
- confirmed the listed license is MIT in both

Thanks for your attention to this and let me know if you need anything else on my end!
