### 2025-09-26 938bbb7
Gaboardi edits -- review pre-accept
