+++ paper/paper.md
@@ -28,21 +28,19 @@ bibliography: paper.bib
 
 # Summary
 
-`Craterpy` is an open-source Python package that reads impact crater databases, defines spatial regions of interest, and extracts data associated with each crater into standard Python data structures. It provides a simple and consistent toolkit for parsing, cleaning, and georeferencing crater data into Geopandas `GeoDataFrames`, and allows visualization and extraction of statistics from georeferenced raster imagery of planetary bodies across the Solar System (Figure 1). `Craterpy` simplifies many error-prone manual tasks like defining crater regions of interest in the correct projection, reprojecting to the format of the desired raster and handling pole-crossing or meridian-crossing geometries. `Craterpy` helps users access and analyze crater statistical data from lists of craters that can number in the millions from planetary images that can be GBs to TBs in size.  All of these features come from a clean and efficient object-oriented user interface, built to be interoperable with the planetary geospatial Python stack and Geographic Information System (GIS) software. 
+`Craterpy` is an open-source Python package that reads impact crater databases, defines spatial regions of interest, and extracts data associated with each crater into standard Python data structures. It provides a simple and consistent toolkit for parsing, cleaning, and georeferencing crater data into Geopandas `GeoDataFrames`, and allows visualization and extraction of statistics from georeferenced raster imagery of planetary bodies across the Solar System (\autoref{fig:vesta}. `Craterpy` simplifies many error-prone manual tasks like defining crater regions of interest in the correct projection, reprojecting to the format of the desired raster and handling pole-crossing or meridian-crossing geometries. `Craterpy` helps users access and analyze crater statistical data from lists of craters that can number in the millions from planetary images that can be GBs to TBs in size.  All of these features come from a clean and efficient object-oriented user interface, built to be interoperable with the planetary geospatial Python stack and Geographic Information System (GIS) software. 
 
-![Plot of a Vesta crater database on a global mosaic, as generated by craterpy v0.9.3. See repository for latest API documentation and worked examples.](../docs/_images/readme_vesta_cdb.png)
+![Plot of a Vesta crater database on a global mosaic, as generated by craterpy v0.9.3. See the repository for latest API documentation and worked examples.\label{fig:vesta}](../docs/_images/readme_vesta_cdb.png)
 
-Impact craters are ubiquitous across the Solar System. For every mission to a new planetary body and every enhancement of data resolution upon repeat visits, planetary scientists generate crater databases, typically in the form of tables of latitude, longitude, and diameter [@robbins2018;@herrick2018;@kirchoff2015;@zeilnhofer2021;@ghent2024;@robbins2012]. Crater and their ejecta, material excavated and deposited around them, reveal information about the age (chronology), layering (stratigraphy), and evolution of planetary surfaces and have been central to numerous planetary science investigations [@baby2024;@singer2025;@parekh2021;@asphaug1996;@li2018a;@venkatraman2023;@blewett2016;@elder2019]. For example, craters have been used to study the rate of physical and chemical alteration of the Moon's surface over time through data analysis of crater ejecta [@grier2001;@taiudovicic2021;@chertok2023;@nypaver2021a]. `Craterpy` provides a user-friendly toolkit to accurately and efficiently extract and process crater image data to facilitate and enable a wide range of planetary geological investigations. 
-
[…]
