### 2025-09-27 #8663 @jGaboardi
@cjtu Once you have reviewed and either accepted or rejected my suggestions in [cjtu/craterpy#102](https://github.com/cjtu/craterpy/pull/102) (it is not compulsory for you to accept), I will continue my post-review checks.
