### 2025-09-29 #8663 @jGaboardi
@cjtu -- While there didn't show up as a problem for the DOIs in initial checking, these below need fixing in `paper.bib`.

```
Element doi: [facet 'pattern'] The value '10/gn8s9c' is not accepted by the pattern '10\.[0-9]{4,9}/.{1,200}'.
Element doi: [facet 'pattern'] The value '10/cczr33' is not accepted by the pattern '10\.[0-9]{4,9}/.{1,200}'.
Element doi: [facet 'pattern'] The value '10/gnbjq7' is not accepted by the pattern '10\.[0-9]{4,9}/.{1,200}'.
Element doi: [facet 'pattern'] The value '10/ghgp7v' is not accepted by the pattern '10\.[0-9]{4,9}/.{1,200}'.
Element doi: [facet 'pattern'] The value '10/gmpfzn' is not accepted by the pattern '10\.[0-9]{4,9}/.{1,200}'.
```

For example, [`10/gn8s9c`](https://github.com/cjtu/craterpy/blob/e370f041ff16a6e5817b41f55b8f63e692fd123b/paper/paper.bib#L82) is not an actual DOI – it should be [`10.1029/2019JE006128`](https://doi.org/10.1029/2019JE006128). Once these 5 entries have been corrected, please ping back here.
