### 2025-09-29 f48147b
Fix DOIs in paper.bib
