+++ paper/paper.bib
@@ -79,7 +79,7 @@
   number = {12},
   pages = {3373--3384},
   issn = {2169-9100},
-  doi = {10/gn8s9c},
+  doi = {10.1029/2019JE006128},
   urldate = {2022-01-24},
   langid = {english},
   keywords = {impact craters,lunar regolith,regolith formation and evolution,rock breakdown,subsurface rocks,surface evolution}
@@ -109,7 +109,7 @@
   number = {E12},
   pages = {32847--32847},
   issn = {0148-0227},
-  doi = {10/cczr33},
+  doi = {10.1029/1999JE001160},
   keywords = {doi:10.102,http://dx.doi.org/10.1029/1999JE001160}
 }
 
@@ -184,7 +184,7 @@
   volume = {126},
   number = {7},
   issn = {2169-9097, 2169-9100},
-  doi = {10/gnbjq7},
+  doi = {10.1029/2021JE006897},
   urldate = {2021-11-03},
   langid = {english}
 }
@@ -228,7 +228,7 @@
   month = nov,
   journal = {Journal of Geophysical Research: Planets},
   publisher = {John Wiley \& Sons, Ltd},
-  doi = {10/ghgp7v},
+  doi = {10.1029/2018JE005592},
   keywords = {Crater ellipticity,Craters,Database comparison,Moon,Secondary craters}
 }
 
@@ -282,7 +282,7 @@
   volume = {48},
   number = {14},
   issn = {0094-8276, 1944-8007},
-  doi = {10/gmpfzn},
+  doi = {10.1029/2020GL092198},
   urldate = {2021-09-03},
   copyright = {All rights reserved},
   langid = {english}
