### 2025-09-29 #8663 @cjtu
Ah I didn't catch these either. Just pushed a fix for these 5 DOIs!
