### 2025-09-29 #8663 @jGaboardi
Nice work on this @cjtu - now this is over to the track editor for further processing.

And thank you @athulpg007 & @marcsit for your timely and thoughtful reviews!
